Refactor tests to use shared cleanup and mock helpers
- Extracted common cleanup logic to
- Extracted shared mocks to
- Updated  and  to use these helpers
